Split validation into separate files to prep for refactoring
