+++ samples/separate/TodaySchema.h
@@ -6,7 +6,7 @@
 #ifndef TODAYSCHEMA_H
 #define TODAYSCHEMA_H
 
-#include <graphqlservice/GraphQLService.h>
+#include "graphqlservice/GraphQLService.h"
 
 #include <memory>
 #include <string>
