+++ samples/today/sample.cpp
@@ -3,7 +3,7 @@
 
 #include "TodayMock.h"
 
-#include <graphqlservice/JSONResponse.h>
+#include "graphqlservice/JSONResponse.h"
 
 #include <iterator>
 #include <iostream>
