+++ samples/validation/ValidationSchema.cpp
@@ -3,7 +3,7 @@
 
 #include "ValidationSchema.h"
 
-#include <graphqlservice/Introspection.h>
+#include "graphqlservice/Introspection.h"
 
 #include <algorithm>
 #include <functional>
