+++ samples/validation/ValidationSchema.h
@@ -6,7 +6,7 @@
 #ifndef VALIDATIONSCHEMA_H
 #define VALIDATIONSCHEMA_H
 
-#include <graphqlservice/GraphQLService.h>
+#include "graphqlservice/GraphQLService.h"
 
 #include <memory>
 #include <string>
