+++ src/GraphQLResponse.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include <graphqlservice/GraphQLResponse.h>
+#include "graphqlservice/GraphQLResponse.h"
 
 #include <stdexcept>
 #include <variant>
