+++ src/Introspection.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include <graphqlservice/Introspection.h>
+#include "graphqlservice/Introspection.h"
 
 namespace graphql::introspection {
